More utilties moved
