+++ SoulsFormats/Formats/DCX.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using SoulsFormats.Compression;
+using System;
 using System.IO;
 using System.IO.Compression;
 
@@ -159,7 +160,7 @@ namespace SoulsFormats
             switch (type)
             {
                 case Type.Zlib:
-                    return SFUtil.ReadZlib(br, (int)br.Length);
+                    return ZlibHelper.ReadZlib(br, (int)br.Length);
                 case Type.DCP_EDGE:
                     return DecompressDCPEDGE(br);
                 case Type.DCP_DFLT:
@@ -202,7 +203,7 @@ namespace SoulsFormats
             int uncompressedSize = br.ReadInt32();
             int compressedSize = br.ReadInt32();
 
-            byte[] decompressed = SFUtil.ReadZlib(br, compressedSize);
+            byte[] decompressed = ZlibHelper.ReadZlib(br, compressedSize);
 
             br.AssertASCII("DCA\0");
             br.AssertInt32(8);
@@ -380,7 +381,7 @@ namespace SoulsFormats
             br.AssertASCII("DCA\0");
             int compressedHeaderLength = br.ReadInt32();
 
-            return SFUtil.ReadZlib(br, Convert.ToInt32(br.Length - br.Position));
+            return ZlibHelper.ReadZlib(br, Convert.ToInt32(br.Length - br.Position));
         }
 
         private static byte[] DecompressDCXKRAK(BinaryReaderEx br, bool maxCompression = false)
@@ -444,9 +445,7 @@ namespace SoulsFormats
             br.AssertASCII("DCA\0");
             br.AssertInt32(8);
 
-            byte[] decompressed = SFUtil.ReadZstd(br, compressedSize);
-
-            return decompressed;
+            return ZstdHelper.ReadZstd(br, compressedSize);
         }
 
         #endregion
@@ -486,7 +485,7 @@ namespace SoulsFormats
             switch (type)
             {
                 case Type.Zlib:
-                    SFUtil.WriteZlib(bw, 0xDA, data);
+                    ZlibHelper.WriteZlib(bw, 0xDA, data);
                     return;
                 case Type.DCP_EDGE:
                     return;
@@ -538,7 +537,7 @@ namespace SoulsFormats
             bw.WriteInt32(data.Length);
             bw.ReserveInt32("CompressedSize");
 
-            int compressedSize = SFUtil.WriteZlib(bw, 0xDA, data);
+            int compressedSize = ZlibHelper.WriteZlib(bw, 0xDA, data);
             bw.FillInt32("CompressedSize", compressedSize);
 
             bw.WriteASCII("DCA\0");
@@ -701,7 +700,7 @@ namespace SoulsFormats
             bw.WriteInt32(8);
 
             long compressedStart = bw.Position;
-            SFUtil.WriteZlib(bw, 0xDA, data);
+            ZlibHelper.WriteZlib(bw, 0xDA, data);
             bw.FillInt32("CompressedSize", (int)(bw.Position - compressedStart));
         }
 
@@ -738,7 +737,7 @@ namespace SoulsFormats
 
         private static void CompressDCXZSTD(byte[] data, BinaryWriterEx bw, int compressionLevel = 15)
         {
-            byte[] compressed = SFUtil.WriteZstd(data, compressionLevel);
+            byte[] compressed = ZstdHelper.WriteZstd(data, compressionLevel);
 
             bw.WriteASCII("DCX\0");
             bw.WriteInt32(0x11000);
