+++ SoulsFormats/Utilities/Cryptography/SL2Decryptor.cs
@@ -1,0 +1,104 @@
+﻿using System.IO;
+using System.Security.Cryptography;
+using System;
+
+namespace SoulsFormats.Cryptography
+{
+    /// <summary>
+    /// Helper methods for decrypting and encrypting SL2 saves.
+    /// </summary>
+    public static class SL2Decryptor
+    {
+        private static readonly byte[] ds2SaveKey = new byte[16] { 0xB7, 0xFD, 0x46, 0x3E, 0x4A, 0x9C, 0x11, 0x02, 0xDF, 0x17, 0x39, 0xE5, 0xF3, 0xB2, 0xA5, 0x0F };
+        private static readonly byte[] scholarSaveKey = new byte[16] { 0x59, 0x9F, 0x9B, 0x69, 0x96, 0x40, 0xA5, 0x52, 0x36, 0xEE, 0x2D, 0x70, 0x83, 0x5E, 0xC7, 0x44 };
+        private static readonly byte[] ds3SaveKey = new byte[16] { 0xFD, 0x46, 0x4D, 0x69, 0x5E, 0x69, 0xA3, 0x9A, 0x10, 0xE3, 0x19, 0xA7, 0xAC, 0xE8, 0xB7, 0xFA };
+
+        /// <summary>
+        /// Returns a copy of the key used for encrypting original DS2 save files on PC.
+        /// </summary>
+        public static byte[] GetDS2SaveKey()
+        {
+            return (byte[])ds2SaveKey.Clone();
+        }
+
+        /// <summary>
+        /// Returns a copy of the key used for encrypting DS2 SotFS save files on PC.
+        /// </summary>
+        public static byte[] GetScholarSaveKey()
+        {
+            return (byte[])scholarSaveKey.Clone();
+        }
+
+        /// <summary>
+        /// Returns a copy of the key used for encrypting DS3 save files on PC.
+        /// </summary>
+        public static byte[] GetDS3SaveKey()
+        {
+            return (byte[])ds3SaveKey.Clone();
+        }
+
+        /// <summary>
+        /// Decrypts a file from a DS2/DS3 SL2. Do not remove the hash and IV before calling.
+        /// </summary>
+        public static byte[] DecryptSL2File(byte[] encrypted, byte[] key)
+        {
+            // Just leaving this here for documentation
+            //byte[] hash = new byte[16];
+            //Buffer.BlockCopy(encrypted, 0, hash, 0, 16);
+
+            byte[] iv = new byte[16];
+            Buffer.BlockCopy(encrypted, 16, iv, 0, 16);
+
+            using (Aes aes = Aes.Create())
+            {
+                aes.Mode = CipherMode.CBC;
+                aes.BlockSize = 128;
+                // PKCS7-style padding is used, but they don't include the minimum padding
+                // so it can't be stripped safely
+                aes.Padding = PaddingMode.None;
+                aes.Key = key;
+                aes.IV = iv;
+
+                ICryptoTransform decryptor = aes.CreateDecryptor();
+                using (var encStream = new MemoryStream(encrypted, 32, encrypted.Length - 32))
+                using (var cryptoStream = new CryptoStream(encStream, decryptor, CryptoStreamMode.Read))
+                using (var decStream = new MemoryStream())
+                {
+                    cryptoStream.CopyTo(decStream);
+                    return decStream.ToArray();
+                }
+            }
+        }
+
+        /// <summary>
+        /// Encrypts a file for a DS2/DS3 SL2. Result includes the hash and IV.
+        /// </summary>
+        public static byte[] EncryptSL2File(byte[] decrypted, byte[] key)
+        {
+            using (Aes aes = Aes.Create())
+            {
+                aes.Mode = CipherMode.CBC;
+                aes.BlockSize = 128;
+                aes.Padding = PaddingMode.None;
+                aes.Key = key;
+                aes.GenerateIV();
+
+                ICryptoTransform encryptor = aes.CreateEncryptor();
+                using (var decStream = new MemoryStream(decrypted))
+                using (var cryptoStream = new CryptoStream(decStream, encryptor, CryptoStreamMode.Read))
+                using (var encStream = new MemoryStream())
+                using (var md5 = MD5.Create())
+                {
+                    encStream.Write(aes.IV, 0, 16);
+                    cryptoStream.CopyTo(encStream);
+                    byte[] encrypted = new byte[encStream.Length + 16];
+                    encStream.Position = 0;
+                    encStream.Read(encrypted, 16, (int)encStream.Length);
+                    byte[] hash = md5.ComputeHash(encrypted, 16, encrypted.Length - 16);
+                    Buffer.BlockCopy(hash, 0, encrypted, 0, 16);
+                    return encrypted;
+                }
+            }
+        }
+    }
+}
